+++ src/components/offer/OfferForm.js
@@ -1,12 +1,10 @@
 import React from 'react';
-import { Layout, Form, Input, Row, Col, Select, Button, DatePicker } from 'antd';
-import ImageUploadAndNameInputs from 'components/inputs/ImageUploadAndNameInputs';
-import SearchFunnel from 'components/inputs/SearchFunnel';
-import TitleDivider from './TitleDivider';
+import { Layout, Form, Input, Row, Col, Select, AutoComplete, Button, DatePicker } from 'antd';
+import { SearchFunnel, ImageUploadAndNameInputs } from 'components/shared';
 
 const { Option } = Select;
 
-const PathwayForm = React.forwardRef((props, ref) => {
+const OfferForm = React.forwardRef((props, ref) => {
     const [ form ] = Form.useForm();
 
     return (
@@ -15,15 +13,44 @@ const PathwayForm = React.forwardRef((props, ref) => {
                 form={form}
             >
                 <ImageUploadAndNameInputs>
-                    <Form.Item
-                        label="Description"
-                        name="description"
-                        labelAlign={"left"}
-                        colon={false}
-                        className="mb-0 inherit"
-                    >
-                        <Input.TextArea rows={2} />
-                    </Form.Item>
+                    <Row gutter={8}>
+                        <Col span={18}>
+                            <Form.Item
+                                label="Provider"
+                                name="provider"
+                                labelAlign={"left"}
+                                colon={false}
+                                className="mb-0 inherit"
+                            >
+                                <Row>
+                                    <Col span={17}>
+                                        <AutoComplete />
+                                    </Col>
+                                    <Col type={5}>
+                                        <Button type="primary">
+                                                Use Image
+                                        </Button>
+                                    </Col>
+                                </Row>
+                            </Form.Item>
+                        </Col>
+                        <Col span={6}>
+                            <Form.Item
+                                label="Generic Offer"
+                                name="generic_offer"
+                                labelAlign={"left"}
+                                colon={false}
+                                className="mb-0 inherit"
+                            >
+                                <Select>
+                                    <Option value="save_the_planet">Save the Planet</Option>
+                                    <Option value="web_development">Web Development</Option>
+                                    <Option value="science_offers">Science Offers</Option>
+                                    <Option value="finance_and_business">Finance &#38; Business</Option>
+                                </Select>
+                            </Form.Item>
+                        </Col>
+                    </Row>
                     <Row gutter={8}>
                         <Col span={10}>
                             <Form.Item
@@ -52,21 +79,21 @@ const PathwayForm = React.forwardRef((props, ref) => {
                         </Col>
                     </Row>
                 </ImageUploadAndNameInputs>
-                <TitleDivider title={"Add Offers Group"} />
-                <Row>
-                    <Col span={8}>
-                        <Input
-                            className="w-full"
-                            placeholder="Group Name"
-                        />
-                    </Col>
-                    <Col type={4}>
-                        <Button type="primary">
-                            Add Group
-                        </Button>
-                    </Col>
-                </Row>
-                <TitleDivider title={"Pathway Offers Groups"} />
+                <Form.Item
+                    label="Description"
+                    name="description"
+                    labelAlign={"left"}
+                    colon={false}
+                    className="mb-0 inherit"
+                >
+                    <Input.TextArea rows={4} />
+                </Form.Item>
+                <SearchFunnel
+                    title={"Related Offers"}
+                />
+                <SearchFunnel
+                    title={"Pre-requisites"}
+                />
                 <SearchFunnel
                     title={"Topics"}
                 />
@@ -88,53 +115,16 @@ const PathwayForm = React.forwardRef((props, ref) => {
                     </Col>
                     <Col span={6}>
                         <Form.Item
-                            label="Generic Type"
-                            name="type"
-                            labelAlign={"left"}
-                            colon={false}
-                            className="mb-0 inherit"
-                        >
-                            <Input />
-                        </Form.Item>
-                    </Col>
-                    <Col span={12}>
-                        <Form.Item
-                            label="Earnings"
-                            name="earnings"
-                            labelAlign={"left"}
-                            colon={false}
-                            className="mb-0 inherit"
-                        >
-                            <Input type="number" />
-                        </Form.Item>
-                    </Col>
-                </Row>
-                <Row gutter={8}>
-                <Col span={6}>
-                        <Form.Item
-                            label="Length"
-                            name="length"
-                            labelAlign={"left"}
-                            colon={false}
-                            className="mb-0 inherit"
-                        >
-                            <Input />
-                        </Form.Item>
-                    </Col>
-                    <Col span={6}>
-                        <Form.Item
-                            label="Length Unit"
-                            name="length_unit"
+                            label="Day &#38; Time"
+                            name="datetime"
                             labelAlign={"left"}
                             colon={false}
                             className="mb-0 inherit"
                         >
                             <Select>
-                                <Option value="weeks">Weeks</Option>
-                                <Option value="biweek">Biweeks</Option>
-                                <Option value="months">Months</Option>
-                                <Option value="quarters">Quarters</Option>
-                                <Option value="semesters">Semesters</Option>
+                                <Option value="morning">Morning</Option>
+                                <Option value="afternoon">Afternoon</Option>
+                                <Option value="evening">Evening</Option>
                             </Select>
                         </Form.Item>
                     </Col>
@@ -168,6 +158,28 @@ const PathwayForm = React.forwardRef((props, ref) => {
                 <Row gutter={8}>
                     <Col span={6}>
                         <Form.Item
+                            label="Cost"
+                            name="cost"
+                            labelAlign={"left"}
+                            colon={false}
+                            className="mb-0 inherit"
+                        >
+                            <Input />
+                        </Form.Item>
+                    </Col>
+                    <Col span={6}>
+                        <Form.Item
+                            label="Cost Unit"
+                            name="cost_unit"
+                            labelAlign={"left"}
+                            colon={false}
+                            className="mb-0 inherit"
+                        >
+                            <Input />
+                        </Form.Item>
+                    </Col>
+                    <Col span={6}>
+                        <Form.Item
                             label="Credit"
                             name="credit"
                             labelAlign={"left"}
@@ -194,6 +206,8 @@ const PathwayForm = React.forwardRef((props, ref) => {
                             </Select>
                         </Form.Item>
                     </Col>
+                </Row>
+                <Row gutter={8}>
                     <Col span={6}>
                         <Form.Item
                             label="Pay"
@@ -222,21 +236,38 @@ const PathwayForm = React.forwardRef((props, ref) => {
                             </Select>
                         </Form.Item>
                     </Col>
+                    <Col span={6}>
+                        <Form.Item
+                            label="Length"
+                            name="length"
+                            labelAlign={"left"}
+                            colon={false}
+                            className="mb-0 inherit"
+                        >
+                            <Input />
+                        </Form.Item>
+                    </Col>
+                    <Col span={6}>
+                        <Form.Item
+                            label="Length Unit"
+                            name="length_unit"
+                            labelAlign={"left"}
+                            colon={false}
+                            className="mb-0 inherit"
+                        >
+                            <Select>
+                                <Option value="weeks">Weeks</Option>
+                                <Option value="biweek">Biweeks</Option>
+                                <Option value="months">Months</Option>
+                                <Option value="quarters">Quarters</Option>
+                                <Option value="semesters">Semesters</Option>
+                            </Select>
+                        </Form.Item>
+                    </Col>
                 </Row>
-                <Col span={12}>
-                    <Form.Item
-                        label="Outlook"
-                        name="outlook"
-                        labelAlign={"left"}
-                        colon={false}
-                        className="mb-0 inherit"
-                    >
-                        <Input />
-                    </Form.Item>
-                </Col>
             </Form>
         </Layout>
     );
 });
 
-export default PathwayForm;
+export default OfferForm;
