+++ src/components/shared/ImageUploadAndNameInputs.js
