+++ src/components/shared/index.js
@@ -1,0 +1,7 @@
+import ImageUploadAndNameInputs from './ImageUploadAndNameInputs';
+import SearchFunnel from './SearchFunnel';
+
+export {
+    ImageUploadAndNameInputs,
+    SearchFunnel,
+};
